+++ ex_webrtc/integration_test/test_videoroom/config/config.exs
@@ -14,7 +14,7 @@ config :esbuild,
   version: "0.12.18",
   default: [
     args:
-      ~w(js/app.js --bundle --target=es2016 --outdir=../priv/static/assets --external:/fonts/* --external:/images/*),
+      ~w(src/index.ts --bundle --target=es2020 --outdir=../priv/static/assets --external:/fonts/* --external:/images/*),
     cd: Path.expand("../assets", __DIR__),
     env: %{"NODE_PATH" => Path.expand("../deps", __DIR__)}
   ]
@@ -25,10 +25,7 @@ config :logger, :console,
 
 config :logger,
   compile_time_purge_matching: [
-    [level_lower_than: :info],
-    # Silence irrelevant warnings caused by resending handshake events
-    [module: Membrane.SRTP.Encryptor, function: "handle_event/4", level_lower_than: :error],
-    [module: Membrane.ICE, level_lower_than: :error]
+    [level_lower_than: :info]
   ]
 
 config :logger, level: :debug