+++ ex_webrtc/lib/ex_webrtc/track_receiver.ex
@@ -36,7 +36,7 @@ defmodule Membrane.RTC.Engine.Endpoint.ExWebRTC.TrackReceiver do
     VoiceActivityChanged
   }
 
-  alias Membrane.RTC.Engine.Endpoint.WebRTC.Metrics
+  alias Membrane.RTC.Engine.Endpoint.ExWebRTC.Metrics
   alias Membrane.RTC.Engine.Track
   alias Membrane.RTCP.SenderReportPacket
   alias Membrane.RTCPEvent