+++ ex_webrtc/test/ex_webrtc/peer_connection_handler_test.exs
@@ -12,6 +12,7 @@ defmodule Membrane.RTC.Engine.Endpoint.ExWebRTC.PeerConnectionHandlerTest do
   alias Membrane.RTC.Engine.Track
 
   @endpoint_id "ex_webrtc_endpoint"
+  @track_metadata %{"server" => %{"displayName" => "mrwebrtc"}, "peer" => %{}}
 
   setup do
     {:ok, pc} = PeerConnection.start_link()
@@ -24,12 +25,13 @@ defmodule Membrane.RTC.Engine.Endpoint.ExWebRTC.PeerConnectionHandlerTest do
   test "peer adds single video track", %{pc: pc, pipeline: pipeline} do
     track_id = UUID.uuid4()
     mid_to_track_id = %{"0" => track_id}
+    track_id_to_metadata = %{track_id => @track_metadata}
     track = MediaStreamTrack.new(:video, [MediaStreamTrack.generate_stream_id()])
     {:ok, _transceiver} = PeerConnection.add_transceiver(pc, track, direction: :sendonly)
     {:ok, offer} = PeerConnection.create_offer(pc)
     :ok = PeerConnection.set_local_description(pc, offer)
 
-    media_event = sdp_offer(offer, mid_to_track_id)
+    media_event = sdp_offer(offer, mid_to_track_id, track_id_to_metadata)
 
     outbound_tracks = %{}
     Pipeline.notify_child(pipeline, :handler, {:offer, media_event, outbound_tracks})
@@ -40,13 +42,19 @@ defmodule Membrane.RTC.Engine.Endpoint.ExWebRTC.PeerConnectionHandlerTest do
       {:answer, %{"type" => "answer", "sdp" => _sdp} = answer, _new_mid_to_track_id}
     )
 
-    assert_pipeline_notified(pipeline, :handler, {:tracks, tracks})
+    assert_pipeline_notified(pipeline, :handler, {:new_tracks, tracks})
 
-    assert length(tracks) == 1
-    track = List.first(tracks)
+    [engine_track] = tracks
 
-    assert %{id: ^track_id, type: :video, origin: @endpoint_id, encoding: :VP8, variants: [:high]} =
-             track
+    assert %{
+             id: ^track_id,
+             type: :video,
+             origin: @endpoint_id,
+             encoding: :VP8,
+             variants: [:high],
+             metadata: @track_metadata
+           } =
+             engine_track
 
     answer = SessionDescription.from_json(answer)
     PeerConnection.set_remote_description(pc, answer)
@@ -74,7 +82,7 @@ defmodule Membrane.RTC.Engine.Endpoint.ExWebRTC.PeerConnectionHandlerTest do
       {:answer, %{"type" => "answer", "sdp" => _sdp} = answer, new_mid_to_track_id}
     )
 
-    refute_pipeline_notified(pipeline, :handler, {:tracks, _tracks})
+    refute_pipeline_notified(pipeline, :handler, {:new_tracks, _tracks})
 
     answer = SessionDescription.from_json(answer)
     PeerConnection.set_remote_description(pc, answer)
@@ -108,7 +116,7 @@ defmodule Membrane.RTC.Engine.Endpoint.ExWebRTC.PeerConnectionHandlerTest do
       {:answer, %{"type" => "answer", "sdp" => _sdp} = answer, _new_mid_to_track_id}
     )
 
-    assert_pipeline_notified(pipeline, :handler, {:tracks, tracks})
+    assert_pipeline_notified(pipeline, :handler, {:new_tracks, tracks})
 
     assert length(tracks) == 2
     engine_audio_track = Enum.find(tracks, &(&1.type == :audio))
@@ -163,7 +171,7 @@ defmodule Membrane.RTC.Engine.Endpoint.ExWebRTC.PeerConnectionHandlerTest do
 
     assert_pipeline_notified(pipeline, :handler, :negotiation_done)
 
-    assert_pipeline_notified(pipeline, :handler, {:tracks_removed, removed_tracks}, 20_000)
+    assert_pipeline_notified(pipeline, :handler, {:tracks_removed, removed_tracks})
     assert removed_tracks |> List.first() == engine_track.id
   end
 
@@ -171,28 +179,33 @@ defmodule Membrane.RTC.Engine.Endpoint.ExWebRTC.PeerConnectionHandlerTest do
     [
       child(:handler, %PeerConnectionHandler{
         endpoint_id: @endpoint_id,
-        ice_port_range: 50_000..50_050
+        ice_port_range: 50_000..50_050,
+        video_codecs: nil
       })
     ]
   end
 
-  defp sdp_offer(offer, mid_to_track_id \\ %{}) do
-    %{sdp_offer: SessionDescription.to_json(offer), mid_to_track_id: mid_to_track_id}
+  defp sdp_offer(offer, mid_to_track_id \\ %{}, track_id_to_metadata \\ %{}) do
+    %{
+      sdp_offer: SessionDescription.to_json(offer),
+      mid_to_track_id: mid_to_track_id,
+      track_id_to_track_metadata: track_id_to_metadata
+    }
   end
 
   defp engine_video_track() do
     codec =
       PeerConnection.Configuration.default_video_codecs()
-      |> Enum.filter(&(&1.mime_type == "video/VP8"))
-      |> List.first()
+      |> Enum.find(&(&1.mime_type == "video/H264"))
 
     Track.new(
       :video,
       Track.stream_id(),
       @endpoint_id,
-      :VP8,
+      :H264,
       codec.clock_rate,
-      codec.sdp_fmtp_line
+      codec.sdp_fmtp_line,
+      id: UUID.uuid4()
     )
   end
 
@@ -209,7 +222,7 @@ defmodule Membrane.RTC.Engine.Endpoint.ExWebRTC.PeerConnectionHandlerTest do
     Pipeline.notify_child(pipeline, :handler, {:offer, media_event, %{}})
 
     assert_pipeline_notified(pipeline, :handler, {:answer, %{"type" => "answer"} = answer, _mids})
-    assert_pipeline_notified(pipeline, :handler, {:tracks, tracks})
+    assert_pipeline_notified(pipeline, :handler, {:new_tracks, tracks})
     [engine_track] = tracks
 
     assert engine_track.id == track_id